refactor(load-test): 테스트 유형별 구조로 분리 (#104)
