+++ scripts/load-test/batch_embedding/__init__.py
@@ -1,0 +1,4 @@
+from batch_embedding.session import BatchEmbeddingSession, BatchRunConfig
+from batch_embedding.target import BatchTarget
+
+__all__ = ["BatchEmbeddingSession", "BatchRunConfig", "BatchTarget"]
