+++ scripts/load-test/batch_embedding/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Synthetic fixture generation and validation for batch embedding tests."""
