+++ scripts/load-test/batch_embedding/fixtures/validate.py
@@ -6,7 +6,7 @@ from collections import Counter
 from pathlib import Path
 from typing import Any
 
-from generate_enriched_text_fixture import (
+from batch_embedding.fixtures.generate import (
     BATCH_SIZE,
     CAPACITY_TOKEN_BUCKETS,
     CONTENT_PROFILES,
@@ -317,7 +317,7 @@ def build_parser(repo_root: Path) -> argparse.ArgumentParser:
 
 def main() -> None:
     script_path = Path(__file__).resolve()
-    repo_root = script_path.parents[2] if len(script_path.parents) > 2 else Path.cwd()
+    repo_root = script_path.parents[4] if len(script_path.parents) > 4 else Path.cwd()
     arguments = build_parser(repo_root).parse_args()
     db_profile = json.loads(arguments.db_profile.read_text(encoding="utf-8"))
     limits = read_limits(db_profile)
