+++ scripts/load-test/search_embedding/__init__.py
@@ -1,0 +1,4 @@
+from search_embedding.session import SearchEmbeddingSession, SearchRunConfig
+from search_embedding.target import SearchTarget
+
+__all__ = ["SearchEmbeddingSession", "SearchRunConfig", "SearchTarget"]
