+++ scripts/load-test/tests/batch/__init__.py
@@ -1,0 +1,1 @@
+"""Batch embedding load-test tests."""
