+++ scripts/load-test/tests/common/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for shared load-test infrastructure."""
