+++ scripts/load-test/tests/search/__init__.py
@@ -1,0 +1,1 @@
+"""Search embedding load-test tests."""
